TablesPage moved to containers
